requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
drf_spectacular==0.25.1
model-bakery==1.10.1
pytest-django==4.5.2
python-dotenv=0.21.1
pytz==2022.7.1
sqlparse==0.4.3
